Change file structure - move views folder from src/components/ to src/
